Conflicts:
	drivers/i2c/busses/Kconfig
	drivers/net/dm9000.c
	drivers/net/wireless/libertas/if_sdio1.c
	drivers/usb/serial/option.c
	net/bluetooth/rfcomm/core.c

+++ fs/libfs.c
@@ -415,7 +415,8 @@ int simple_write_end(struct file *file, struct address_space *mapping,
  * unique inode values later for this filesystem, then you must take care
  * to pass it an appropriate max_reserved value to avoid collisions.
  */
-int simple_fill_super(struct super_block *s, int magic, struct tree_descr *files)
+int simple_fill_super(struct super_block *s, unsigned long magic,
+		      struct tree_descr *files)
 {
 	struct inode *inode;
 	struct dentry *root;
